+++ lua/plugins/mason.lua
@@ -2,6 +2,7 @@ return {
   {
     "mason-org/mason.nvim",
     opts = {
+      PATH = "append",
       ui = { border = "rounded" },
     },
   },